[feature] Implemented design of widget
