+++ app/src/main/res/xml/small_sunset_hour_widget_info.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
-    android:description="@string/app_name"
-    android:minWidth="250dp"
-    android:minHeight="80dp"
-    android:resizeMode="horizontal|vertical"
+    android:description="@string/small_hour_widget_description"
+    android:minWidth="130dp"
+    android:resizeMode="none"
     android:widgetCategory="home_screen" />
